Distinguish true MP V5, v6 implmentations vs replication of matlab bugs
The matlab implementation of V5 contained a few bugs.  These were corrected in haphazard fashion over time. At least
one discrepancy remained in all production, since 2015.  This discrepancy regards eqn 57, 58.  The estimate
of the proportion female is supposed to use (Ef+ Em) weights, but had always been done using equal weights.

An argument, reproduce.matlab (default FALSE) is used to replicate the MATLAB implimentation.  When FALSE,
the true V5 method is followed.  For V6, the procedure for smooth estimation of proportion female for Total
is the same as in V5.  For MP > 6, two improvements are proposed and mentioned in ltperBoth_AxN code.