CountryLong name made independent of lookup table country.lookup
CAB: The long country name was set inside Write_* functions by table lookup
from package data object country.lookup[].  This meant that new country
names not present in country.lookup would be empty and adding a country
meant modifying and rebuilding the package.  RunHMDCountry() modified to
accept the country descriptor as an argument and pass it to all writing
functions.