bug fix for writing Births from either IDB, LDB and warning of coverage differences
The LDB contains birth information, but for a possible subset of the
data contained in the IDB Births file.  Output the greater set of data
and warn of years where the IDB and LDB do not match.  For some
aggregate countries (GBRNP) the birth series is missing from the IDB and
so the LDB variant is used.